[WIP] Reverted commands structure
