+++ bot/src/handler/commands/start.py
@@ -1,0 +1,8 @@
+from . import Base
+
+
+class Start(Base):
+    name = 'start'
+
+    def execute(self, command):
+        pass
